+++ rotkehlchen/tests/unit/decoders/test_makerdao.py
@@ -1,9 +1,9 @@
 import pytest
 
 from rotkehlchen.chain.ethereum.modules.makerdao.constants import CPT_VAULT
-from rotkehlchen.chain.evm.decoding.constants import CPT_GAS, CPT_SDAI
+from rotkehlchen.chain.evm.decoding.constants import CPT_GAS
 from rotkehlchen.chain.evm.types import string_to_evm_address
-from rotkehlchen.constants.assets import A_DAI, A_ETH, A_SDAI, A_WBTC
+from rotkehlchen.constants.assets import A_ETH, A_WBTC
 from rotkehlchen.fval import FVal
 from rotkehlchen.history.events.structures.evm_event import EvmEvent
 from rotkehlchen.history.events.structures.types import HistoryEventSubType, HistoryEventType
@@ -80,104 +80,5 @@ def test_withdraw_with_transfer_after(ethereum_inquirer, ethereum_accounts):
             counterparty=CPT_VAULT,
             address=string_to_evm_address('0xa4aF35efa3b56d0C4F864B6484Af0C5FCD0B13e2'),
             extra_data={'vault_type': 'WBTC-A'},
-        ),
-    ]
-
-
-@pytest.mark.vcr
-@pytest.mark.parametrize('ethereum_accounts', [['0xa217BDa86b0EDb86eE7d4D6e34F493eDF1ea4F29']])
-def test_withdraw_dai_from_sdai(ethereum_inquirer, ethereum_accounts):
-    tx_hash = deserialize_evm_tx_hash('0x6b2a1f836cfc7c28002e4ac60297daa6d79fcde892d9c3b9ca723dea2f21af5c')  # noqa: E501
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hash)
-    user_address, timestamp, sdai = ethereum_accounts[0], TimestampMS(1695854591000), A_SDAI.resolve_to_evm_token()  # noqa: E501
-    assert events == [
-        EvmEvent(
-            sequence_index=0,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.SPEND,
-            event_subtype=HistoryEventSubType.FEE,
-            asset=A_ETH,
-            amount=FVal('0.001301015216220134'),
-            location_label=user_address,
-            notes='Burn 0.001301015216220134 ETH for gas',
-            tx_hash=tx_hash,
-            counterparty=CPT_GAS,
-        ), EvmEvent(
-            sequence_index=1,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.DEPOSIT,
-            event_subtype=HistoryEventSubType.DEPOSIT_ASSET,
-            asset=A_SDAI,
-            amount=FVal('16.020774067834506624'),
-            location_label=user_address,
-            notes='Return 16.020774067834506624 sDAI to sDAI contract',
-            tx_hash=tx_hash,
-            counterparty=CPT_SDAI,
-            address=sdai.evm_address,
-        ), EvmEvent(
-            sequence_index=2,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.WITHDRAWAL,
-            event_subtype=HistoryEventSubType.REMOVE_ASSET,
-            asset=A_DAI,
-            amount=FVal('16.601085935411927527'),
-            location_label=user_address,
-            notes='Withdraw 16.601085935411927527 DAI from sDAI contract',
-            tx_hash=tx_hash,
-            counterparty=CPT_SDAI,
-            address=sdai.evm_address,
-        ),
-    ]
-
-
-@pytest.mark.vcr
-@pytest.mark.parametrize('ethereum_accounts', [['0xa217BDa86b0EDb86eE7d4D6e34F493eDF1ea4F29']])
-def test_deposit_dai_to_sdai(ethereum_inquirer, ethereum_accounts):
-    user_address = ethereum_accounts[0]
-    tx_hash = deserialize_evm_tx_hash('0x27bd72a2ccd999a44c2a7aaed9090572f34045d62e153362a34715a70ca7a6a7')  # noqa: E501
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=tx_hash)
-    timestamp, address = TimestampMS(1695089927000), A_SDAI.resolve_to_evm_token().evm_address
-    assert events == [
-        EvmEvent(
-            sequence_index=0,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.SPEND,
-            event_subtype=HistoryEventSubType.FEE,
-            asset=A_ETH,
-            amount=FVal(0.00152049387145495),
-            location_label=user_address,
-            notes='Burn 0.00152049387145495 ETH for gas',
-            tx_hash=tx_hash,
-            counterparty=CPT_GAS,
-        ), EvmEvent(
-            sequence_index=1,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.DEPOSIT,
-            event_subtype=HistoryEventSubType.DEPOSIT_ASSET,
-            asset=A_DAI,
-            amount=FVal(16.58145794),
-            location_label=user_address,
-            notes='Deposit 16.58145794 DAI to sDAI contract',
-            tx_hash=tx_hash,
-            counterparty=CPT_SDAI,
-            address=address,
-        ), EvmEvent(
-            sequence_index=2,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.WITHDRAWAL,
-            event_subtype=HistoryEventSubType.REMOVE_ASSET,
-            asset=A_SDAI,
-            amount=FVal('16.020774067834506624'),
-            location_label=user_address,
-            notes='Withdraw 16.020774067834506624 sDAI from sDAI contract',
-            tx_hash=tx_hash,
-            counterparty=CPT_SDAI,
-            address=address,
         ),
     ]